+++ src/cli/src/repl.rs
@@ -46,7 +46,7 @@ impl Repl {
 
         let koto = Koto::with_settings(koto_settings);
 
-        let mut prelude = koto.prelude();
+        let prelude = koto.prelude();
         prelude.add_map("json", koto_json::make_module());
         prelude.add_map("random", koto_random::make_module());
         prelude.add_map("tempfile", koto_tempfile::make_module());