+++ src/koto/benches/koto_benchmark.rs
@@ -10,10 +10,11 @@ static ALLOC: jemallocator::Jemalloc = jemallocator::Jemalloc;
 
 struct BenchmarkRunner {
     runtime: Koto,
+    args: Vec<String>,
 }
 
 impl BenchmarkRunner {
-    fn new(script_path: &str, args: &[String]) -> Self {
+    fn setup(script_path: &str, args: &[String]) -> Self {
         let mut path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
         path.push("..");
         path.push("..");
@@ -25,22 +26,23 @@ impl BenchmarkRunner {
         let mut runtime = Koto::new();
         match runtime.compile(&script) {
             Ok(_) => {
-                runtime.settings.run_tests = true;
-
                 if let Err(error) = runtime.run_with_args(args) {
                     panic!("{}", error);
                 }
-
-                runtime.settings.run_tests = false;
             }
             Err(error) => panic!("{}", error),
         }
 
-        Self { runtime }
+        runtime.set_run_tests(false);
+
+        Self {
+            runtime,
+            args: args.to_vec(),
+        }
     }
 
     fn run(&mut self) {
-        if let Err(error) = self.runtime.run() {
+        if let Err(error) = self.runtime.run_with_args(&self.args) {
             panic!("{}", error);
         }
     }
@@ -48,25 +50,25 @@ impl BenchmarkRunner {
 
 pub fn koto_benchmark(c: &mut Criterion) {
     c.bench_function("fib", |b| {
-        let mut runner = BenchmarkRunner::new("fib_recursive.koto", &[]);
+        let mut runner = BenchmarkRunner::setup("fib_recursive.koto", &[]);
         b.iter(|| {
             runner.run();
         })
     });
     c.bench_function("num4", |b| {
-        let mut runner = BenchmarkRunner::new("num4.koto", &[]);
+        let mut runner = BenchmarkRunner::setup("num4.koto", &[]);
         b.iter(|| {
             runner.run();
         })
     });
     c.bench_function("enumerate", |b| {
-        let mut runner = BenchmarkRunner::new("enumerate.koto", &[]);
+        let mut runner = BenchmarkRunner::setup("enumerate.koto", &[]);
         b.iter(|| {
             runner.run();
         })
     });
     c.bench_function("string_formatting", |b| {
-        let mut runner = BenchmarkRunner::new(
+        let mut runner = BenchmarkRunner::setup(
             "string_formatting.koto",
             &["70".to_string(), "quiet".to_string()],
         );
@@ -75,7 +77,7 @@ pub fn koto_benchmark(c: &mut Criterion) {
         })
     });
     c.bench_function("spectral_norm", |b| {
-        let mut runner = BenchmarkRunner::new(
+        let mut runner = BenchmarkRunner::setup(
             "spectral_norm.koto",
             &["2".to_string(), "quiet".to_string()],
         );
@@ -85,14 +87,14 @@ pub fn koto_benchmark(c: &mut Criterion) {
     });
     c.bench_function("fannkuch", |b| {
         let mut runner =
-            BenchmarkRunner::new("fannkuch.koto", &["4".to_string(), "quiet".to_string()]);
+            BenchmarkRunner::setup("fannkuch.koto", &["4".to_string(), "quiet".to_string()]);
         b.iter(|| {
             runner.run();
         })
     });
     c.bench_function("n_body", |b| {
         let mut runner =
-            BenchmarkRunner::new("n_body.koto", &["10".to_string(), "quiet".to_string()]);
+            BenchmarkRunner::setup("n_body.koto", &["10".to_string(), "quiet".to_string()]);
         b.iter(|| {
             runner.run();
         })