+++ src/runtime/src/core/koto.rs
@@ -3,13 +3,11 @@ use crate::{unexpected_type_error_with_slice, Value, ValueMap, ValueTuple};
 pub fn make_module() -> ValueMap {
     use Value::*;
 
-    let mut result = ValueMap::new();
+    let result = ValueMap::new();
 
     result.add_value("args", Tuple(ValueTuple::default()));
 
-    result.add_fn("exports", |vm, _| {
-        Ok(Value::Map(vm.context_mut().exports.clone()))
-    });
+    result.add_fn("exports", |vm, _| Ok(Value::Map(vm.exports().clone())));
 
     result.add_value("script_dir", Empty);
     result.add_value("script_path", Empty);