+++ src/runtime/src/core/map.rs
@@ -10,7 +10,7 @@ use {
 pub fn make_module() -> ValueMap {
     use Value::*;
 
-    let mut result = ValueMap::new();
+    let result = ValueMap::new();
 
     result.add_fn("clear", |vm, args| match vm.get_args(args) {
         [Map(m)] => {